+++ src/main/java/romever/scan/oasisscan/repository/DebondingRepository.java
@@ -10,7 +10,7 @@ import romever.scan.oasisscan.entity.Debonding;
 import java.util.List;
 
 @Repository
-public interface DebondingRepository extends JpaRepository<Debonding, Integer>,
+public interface DebondingRepository extends JpaRepository<Debonding, Long>,
         JpaSpecificationExecutor<Debonding> {
 
     @Query(value = "select convert(sum(shares),decimal(22,0)) from debonding where delegator=?1", nativeQuery = true)
